requirements.txt
pandas
numpy
scikit-learn
tensorflow
torch
textblob
seaborn
matplotlib
scipy
statsmodels
pyyaml
